+++ DependencyInversionPrinciple/Database.java
@@ -1,0 +1,10 @@
+package DependencyInversionPrinciple;
+
+public class Database implements DatabaseAccess{
+
+    @Override
+    public void save(Employee employee) {
+        // Code to save employee to the database
+    }
+
+}
