+++ DependencyInversionPrinciple/DatabaseAccess.java
@@ -1,0 +1,5 @@
+package DependencyInversionPrinciple;
+
+public interface DatabaseAccess {
+    public void save(Employee employee);   
+}
