+++ DependencyInversionPrinciple/Employee.java
@@ -1,0 +1,13 @@
+package DependencyInversionPrinciple;
+
+public class Employee {
+    private String name;
+    private String address;
+
+    public Employee(String name, String address) {
+        this.name = name;
+        this.address = address;
+    }
+
+    // Getters and setters
+}
