+++ DependencyInversionPrinciple/EmployeeService.java
@@ -1,30 +1,13 @@
 package DependencyInversionPrinciple;
 
-//Apply the Dependency Inversion Principle (DIP) by refactoring a set of classes
-// Note: Also mention why this code violating DIP
-
 public class EmployeeService {
-    private Database database = new Database();
+    private DatabaseAccess databaseAccess;
+
+    public EmployeeService(DatabaseAccess databaseAccess) {
+        this.databaseAccess = databaseAccess;
+    }
 
     public void saveEmployee(Employee employee) {
-        database.save(employee);
+        databaseAccess.save(employee);
     }
-}
-
-public class Database {
-    public void save(Employee employee) {
-        // Code to save employee to the database
-    }
-}
-
-public class Employee {
-    private String name;
-    private String address;
-
-    public Employee(String name, String address) {
-        this.name = name;
-        this.address = address;
-    }
-
-    // Getters and setters
 }
